+++ CHANGELOG.md
@@ -8,6 +8,18 @@ This document contains a summary of changes for this project
 
 ### Removed
 
+## [0.1.1] 02-27-2022
+### Added
+- Basic UI
+- Ability to add course information and display all entered information
+- JavaFX plugin
+- GSON dependency
+- License management plugin
+### Changed
+- Nothing changed
+### Removed
+- Nothing removed
+
 ## [0.0.1] 02-27-2022
 ### Added
 - CHANGELOG.md for a summary of changes